codec refactor - number impl refactor
