+++ elrond-codec/src/impl_for_types/mod.rs
@@ -1,20 +1,15 @@
 mod impl_array;
 mod impl_bool;
 mod impl_bytes;
-mod impl_decode_num_signed;
-mod impl_decode_num_unsigned;
-mod impl_dep_encode_from_no_err;
 mod impl_empty;
-mod impl_encode_num_mimic;
-mod impl_encode_num_signed;
-mod impl_encode_num_unsigned;
 mod impl_non_zero_usize;
-mod impl_numbers;
+mod impl_num_signed;
+mod impl_num_unsigned;
 mod impl_option;
 mod impl_ref;
 mod impl_slice;
 mod impl_string;
-mod impl_top_encode_from_no_err;
 mod impl_tuple;
 mod impl_unit;
 mod impl_vec;
+mod local_macro;
